# http/cur_swagger.yml
openapi: "3.0.0"
info:
  title: Influx API Service
  version: 0.1.0
servers:
  - url: "http://localhost:9999/api/v2"
paths:
...
  /query/ast:
    post:
      description: analyzes flux query and generates a query specification.
      tags:
        - Query
      parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
            - application/json
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
      requestBody:
        description: analyzed flux query to generate abstract syntax tree.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageRequest"
      responses:
        '200':
          description: Abstract syntax tree of flux query.
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /query/spec:
    post:
      description: analyzes flux query and generates a query specification.
      tags:
        - Query
      parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
            - application/json
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
      requestBody:
        description: analyzed flux query to generate specification.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageRequest"
      responses:
        '200':
          description: Specification of flux query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuerySpecification"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /query/plan:
    post:
      description: analyzes flux query or specification and generates a query plan.
      tags:
        - Query
      parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
            - application/json
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
      requestBody:
        description: flux query or specification to generate plan.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanRequest"
      responses:
        '200':
          description: Logical plan of flux query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /query/suggestions:
    get:
      tags:
        - Query
      responses:
        '200':
          description: Suggestions for next functions in call chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FluxSuggestions"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /query/suggestions/{name}:
    parameters:
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: name of branching suggestion
    get:
      tags:
        - Query
      responses:
        '200':
          description: Suggestions for next functions in call chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FluxSuggestions"
        default:
          description: Any response other than 200 is an internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /query/analyze:
   post:
    tags:
      - Query
    summary: analyze an influxql or flux query
    parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
            - application/json
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
    requestBody:
        description: flux or influxql query to analyze
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"
    responses:
        '200':
          description: query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeQueryResponse"
        default:
          description: internal server error
          headers:
            X-Influx-Error:
              description: error string describing the problem
              schema:
                type: string
            X-Influx-Reference:
              description: reference code unique to the error type
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /query:
   get:
    tags:
      - Query
    summary: query influx with specified return formatting. The spec and query fields are mutually exclusive.
    parameters:
      - in: query
        name: org
        description: specifies the organization of the resource
        required: true
        schema:
          type: string
      - in: query
        name: query
        description: query script to execute.
        required: true
        schema:
          type: string
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
    responses:
        '200':
          description: query results
          content:
            text/csv:
              schema:
                type: string
                example: >
                  result,table,_start,_stop,_time,region,host,_value
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:00Z,east,A,15.43
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:20Z,east,B,59.25
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:40Z,east,C,52.62
        '400':
          description: error processing query
          headers:
            X-Influx-Error:
              description: error string describing the problem
              schema:
                type: string
            X-Influx-Reference:
              description: reference code unique to the error type
              schema:
                type: integer
          content:
            text/csv:
              schema:
                type: string
                example: >
                  error,reference
                  Failed to parse query,897
        default:
          description: internal server error
          headers:
            X-Influx-Error:
              description: error string describing the problem
              schema:
                type: string
            X-Influx-Reference:
              description: reference code unique to the error type
              schema:
                type: integer
          content:
            text/csv:
              schema:
                type: string
                example: >
                  error,reference
                  Failed to parse query,897
   post:
    tags:
      - Query
    summary: query an influx
    parameters:
      - in: header
        name: Accept
        description: specifies the return content format. Each response content type will have its own dialect options.
        schema:
          type: string
          description: return format of either CSV or Arrow buffers
          default: text/csv
          enum:
            - text/csv
            - application/vnd.influx.arrow
      - in: header
        name: Content-Type
        schema:
          type: string
          enum:
            - application/json
      - in: header
        name: Authorization
        description: the authorization header should be in the format of `Token <key>`
        schema:
          type: string
      - in: query
        name: org
        description: specifies the name of the organization executing the query.
        schema:
          type: string
    requestBody:
        description: flux query or specification to execute
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"
    responses:
        '200':
          description: query results
          content:
            text/csv:
              schema:
                type: string
                example: >
                  result,table,_start,_stop,_time,region,host,_value
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:00Z,east,A,15.43
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:20Z,east,B,59.25
                  mean,0,2018-05-08T20:50:00Z,2018-05-08T20:51:00Z,2018-05-08T20:50:40Z,east,C,52.62
            application/vnd.influx.arrow:
              schema:
                type: string
                format: binary
        '400':
          description: error processing query
          headers:
            X-Influx-Error:
              description: error string describing the problem
              schema:
                type: string
            X-Influx-Reference:
              description: reference code unique to the error type
              schema:
                type: integer
          content:
            text/csv:
              schema:
                type: string
                example: >
                  error,reference
                  Failed to parse query,897
            application/vnd.influx.arrow:
              schema:
                  type: string
                  format: binary
        default:
          description: internal server error
          headers:
            X-Influx-Error:
              description: error string describing the problem
              schema:
                type: string
            X-Influx-Reference:
              description: reference code unique to the error type
              schema:
                type: integer
          content:
            text/csv:
              schema:
                type: string
                example: >
                  error,reference
                  Failed to parse query,897
            application/vnd.influx.arrow:
              schema:
                  type: string
                  format: binary
...
components:
  schemas:
    LanguageRequest:
      description: flux query to be analyzed.
      type: object
      required:
        - query
      properties:
        query:
          description: flux query script to be analyzed
          type: string
    PlanRequest:
      description: flux query or specification to be planned. The spec and query fields are mutually exclusive.
      type: object
      properties:
        query:
          description: flux query script to be analyzed
          type: string
        spec:
          $ref: "#/components/schemas/QuerySpecification"
    PlanResponse:
      description: flux query or specification to be planned.
      type: object
      required:
        - query
      properties:
        spec:
          $ref: "#/components/schemas/QuerySpecification"
          readOnly: true
        logical:
          description: logical plan of the query.
          readOnly: true
          type: object #TODO(goller): document the logical plan format
        physical:
          description: physical plan of the query.
          readOnly: true
          type: object #TODO(goller): document the physical plan format
    Query:
      description: query influx with specified return formatting. The spec and query fields are mutually exclusive.
      type: object
      required:
        - query
      properties:
        query:
          description: query script to execute.
          type: string
        spec:
          $ref: "#/components/schemas/QuerySpecification"
        type:
          description: type of query
          type: string
          default: flux
          enum:
            - flux
            - influxql
        db:
          description: required for influxql type queries
          type: string
        rp:
          description: required for influxql type queries
          type: string
        cluster:
          description: required for influxql type queries
          type: string
        dialect:
          $ref: "#/components/schemas/Dialect"
    QuerySpecification:
      description: consists of a set of operations and a set of edges between those operations to instruct the query engine to operate.
      type: object
      properties:
        operations:
          type: array
          items:
            type: object
            properties:
              kind:
                description: name of the operation to perform
                type: string
              id:
                description: identifier for this operation; it must be unique per query specification; used in edges
                type: string
              spec:
                description: set of properties that specify details of the operation. These vary by the kind of operation.
                type: object
        edges:
          description: list of declaring a parent child id relationship between operations
          type: array
          items:
            type: object
            properties:
              parent:
                description: id of parent node of child within graph of opertions
                type: string
              child:
                description: id of child node of parent within the graph of operations
                type: string
        resources:
          description: optional set of contraints on the resources the query can consume
          type: object
          properties:
            priority:
              description: priority of the query
              oneOf:
              - type: string
                description: lower value will move to the front of the priority queue
                pattern: '^\d+$'
              - type: string
                description: constants to represent the extreme high and low priorities; high is effectively 0.
                enum:
                  - high
                  - low
            concurrency_quota:
              description: number of concurrent workers allowed to process this query; 0 indicates the planner can pick the optimal concurrency.
              type: integer
              default: 0
            memory_bytes_quota:
              description: number of bytes of RAM this query may consume; 0 means unlimited.
              type: integer
              default: 0
        dialect:
          $ref: "#/components/schemas/Dialect"
    Dialect:
          description: dialect are options to change the default CSV output format; https://www.w3.org/TR/2015/REC-tabular-metadata-20151217/#dialect-descriptions
          type: object
          properties:
            header:
              description: if true, the results will contain a header row
              type: boolean
              default: true
            delimiter:
              description: separator between cells; the default is ,
              type: string
              default: ","
              maxLength: 1
              minLength: 1
            annotations:
              description: https://www.w3.org/TR/2015/REC-tabular-data-model-20151217/#columns
              type: array
              items:
                type: string
                enum:
                  - "group"
                  - "datatype"
                  - "default"
                uniqueItems: true
            commentPrefix:
              description: character prefixed to comment strings
              type: string
              default: "#"
              maxLength: 1
              minLength: 0
            dateTimeFormat:
              description: format of timestamps
              type: string
              default: "RFC3339"
              enum:
                - RFC3339
                - RFC3339Nano
...
    FluxSuggestions:
      type: object
      properties:
        funcs:
          type: object
          properties:
            name:
              type: string
            params:
              type: object
...
    Error:
      properties:
        code:
          description: code is the machine-readable error code.
          readOnly: true
          type: string
          enum:
            - internal error
            - not found
            - conflict
            - invalid
            - empty value
        message:
          readOnly: true
          description: message is a human-readable message.
          type: string
        op:
          readOnly: true
          description: op describes the logical code operation during error. Useful for debugging.
          type: string
        err:
          readOnly: true
          description: err is a stack of errors that occurred during processing of the request. Useful for debugging.
          type: string
      required: [code, message, op, err]
...
    AnalyzeQueryResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              line:
                type: integer
              column:
                type: integer
              character:
                type: integer
              message:
                type: string
...